Add files via upload
Missing a max and min distances for agents due to being a float and not iterable , but will attempt to rectify this. Nested loop was the solution for printing the distances between all agents and the != operater was used to stop agents measuring distances if they were the same agent.